fix(poi): handled no poi inserted case
