+++ client/src/trip/components/Trip.js
@@ -17,7 +17,6 @@ import { useNavigate, Route, Routes } from "react-router-dom";
 import TripHandler from "./TripHandler";
 import TripDetails from "./TripDetails";
 import "./styles.css";
-
 import TripExpViewer from "./TripExpViewer";
 
 export default function Trip() {
@@ -88,7 +87,7 @@ export default function Trip() {
         <div className="card-container">
           {trips.map((trip) => (
             <div className="trip-item" key={trip.TripId}>
-              <h3 className="trip-title">TRIP - {trip.Title}</h3>
+              <h3 className="trip-title">TRIP - {trip.Title} <img src={`https://flaglog.com/codes/official-ratio-120px/${trip.Countries}.png`} style={{ width: "40px"}} /></h3>
               <button onClick={() => showTripDetails(trip.TripId)}>
                 {visible === trip.TripId ? "Hide Details" : "View Details"}
               </button>
@@ -111,7 +110,7 @@ export default function Trip() {
         <div className="card-container">
           {publicTrips.map((trip) => (
             <div className="trip-item" key={trip.TripId}>
-              <h3 className="trip-title">TRIP - {trip.Title}</h3>
+              <h3 className="trip-title">TRIP - {trip.Title} <img src={`https://flaglog.com/codes/official-ratio-120px/${trip.Countries}.png`} style={{ width: "40px"}} /></h3>
               <button onClick={() => showTripDetails(trip.TripId)}>
                 {visible === trip.TripId ? "Hide Details" : "View Details"}
               </button>
